creating dedicated module for viewer base classes
