+++ src/pymodaq/utils/plotting/data_viewers/viewer.py
@@ -1,28 +1,19 @@
-from typing import List, Union, TYPE_CHECKING, Iterable
+from typing import List
 import numpy as np
 
 from qtpy import QtWidgets
-from qtpy.QtCore import QObject, Signal, Slot
-
-from pyqtgraph.graphicsItems import InfiniteLine, ROI
 
 from pymodaq.utils.logger import set_logger, get_module_name
-from pymodaq.utils.data import DataToExport, DataRaw, DataWithAxes, Axis, DataSource
-from pymodaq.utils.exceptions import ViewerError
+from pymodaq.utils.data import DataToExport, DataWithAxes, Axis, DataSource
+
 from pymodaq.utils.enums import BaseEnum, enum_checker
 from pymodaq.utils.factory import ObjectFactory, BuilderBase
 from pymodaq.utils.plotting import data_viewers
 from pymodaq.utils.gui_utils import DockArea, Dock
 from pymodaq.utils.managers.parameter_manager import ParameterManager
 
-if TYPE_CHECKING:
-    from pymodaq.utils.plotting.data_viewers.viewer0D import Viewer0D
-    from pymodaq.utils.plotting.data_viewers.viewer1D import Viewer1D
-    from pymodaq.utils.plotting.data_viewers.viewer2D import Viewer2D
-    from pymodaq.utils.plotting.data_viewers.viewerND import ViewerND
 
-config_viewers = {
-}
+config_viewers = {}
 
 logger = set_logger(get_module_name(__file__))
 
@@ -103,163 +94,6 @@ def create_viewerND(parent: QtWidgets.QWidget, **_ignored):
 
 viewer_factory = ViewerFactory()
 
-
-class ViewerBase(QObject):
-    """Base Class for data viewers implementing all common functionalities
-
-    Parameters
-    ----------
-    parent: QtWidgets.QWidget
-    title: str
-
-    Attributes
-    ----------
-    view: QObject
-        Ui interface of the viewer
-
-    data_to_export_signal: Signal[DataToExport]
-    ROI_changed: Signal
-    crosshair_dragged: Signal[float, float]
-    crosshair_clicked: Signal[bool]
-    sig_double_clicked: Signal[float, float]
-    status_signal: Signal[str]
-    """
-    data_to_export_signal = Signal(DataToExport)
-    _data_to_show_signal = Signal(DataWithAxes)
-
-    ROI_changed = Signal()
-    crosshair_dragged = Signal(float, float)  # Crosshair position in units of scaled top/right axes
-    status_signal = Signal(str)
-    crosshair_clicked = Signal(bool)
-    sig_double_clicked = Signal(float, float)
-
-    def __init__(self, parent: QtWidgets.QWidget = None, title=''):
-        super().__init__()
-        self.title = title if title != '' else self.__class__.__name__
-
-        self._raw_data = None
-        self.data_to_export: DataToExport = DataToExport(name=self.title)
-        self.view: Union[Viewer0D, Viewer1D, Viewer2D, ViewerND] = None
-
-        if parent is None:
-            parent = QtWidgets.QWidget()
-            parent.show()
-        self.parent = parent
-
-        self._display_temporary = False
-
-    @property
-    def has_action(self):
-        """Convenience method"""
-        if hasattr(self.view, 'has_action'):
-            return self.view.has_action
-
-    @property
-    def is_action_checked(self):
-        """Convenience method"""
-        if hasattr(self.view, 'is_action_checked'):
-            return self.view.is_action_checked
-
-    @property
-    def is_action_visible(self):
-        """Convenience method"""
-        if hasattr(self.view, 'is_action_visible'):
-            return self.view.is_action_visible
-
-    @property
-    def set_action_checked(self):
-        """Convenience method"""
-        if hasattr(self.view, 'set_action_checked'):
-            return self.view.set_action_checked
-
-    @property
-    def set_action_visible(self):
-        """Convenience method"""
-        if hasattr(self.view, 'set_action_visible'):
-            return self.view.set_action_visible
-
-    @property
-    def get_action(self):
-        """Convenience method"""
-        if hasattr(self.view, 'get_action'):
-            return self.view.get_action
-
-    @property
-    def toolbar(self):
-        """Convenience property"""
-        if hasattr(self.view, 'toolbar'):
-            return self.view.toolbar
-
-    @property
-    def viewer_type(self):
-        """str: the viewer data type see DATA_TYPES"""
-        return ViewersEnum[self.__class__.__name__].value
-
-    def show_data(self, data: DataWithAxes, **kwargs):
-        """Entrypoint to display data into the viewer
-
-        Parameters
-        ----------
-        data: data_mod.DataFromPlugins
-        """
-        if len(data.shape) > 4:
-            raise ViewerError(f'Ndarray of dim: {len(data.shape)} cannot be plotted using a {self.viewer_type}')
-
-        self.data_to_export = DataToExport(name=self.title)
-        self._raw_data = data
-
-        self._display_temporary = False
-
-        self._show_data(data, **kwargs)
-
-    def show_data_temp(self, data: DataRaw, **kwargs):
-        """Entrypoint to display temporary data into the viewer
-
-        No processed data signal is emitted from the viewer
-
-        Parameters
-        ----------
-        data: data_mod.DataFromPlugins
-        """
-        self._display_temporary = True
-        self.show_data(data, **kwargs)
-
-    def _show_data(self, data: DataRaw):
-        """Specific viewers should implement it"""
-        raise NotImplementedError
-
-    def add_attributes_from_view(self):
-        """Convenience function to add attributes from the view to self"""
-        for attribute in self.convenience_attributes:
-            if hasattr(self.view, attribute):
-                setattr(self, attribute, getattr(self.view, attribute))
-
-    def trigger_action(self, action_name: str):
-        """Convenience function to trigger programmatically one of the action of the related view"""
-        if self.has_action(action_name):
-            self.get_action(action_name).trigger()
-
-    def activate_roi(self, activate=True):
-        """Activate the Roi manager using the corresponding action"""
-        raise NotImplementedError
-
-    def setVisible(self, show=True):
-        """convenience method to show or hide the paretn widget"""
-        self.parent.setVisible(show)
-
-    @property
-    def roi_target(self) -> Union[InfiniteLine.InfiniteLine, ROI.ROI]:
-        """To be implemented if necessary (Viewer1D and above)"""
-        return None
-
-    def move_roi_target(self, pos: Iterable[float] = None, **kwargs):
-        """move a specific read only ROI at the given position on the viewer"""
-        ...
-
-    def show_roi_target(self, show=True):
-        """Show/Hide a specific read only ROI"""
-        if self.roi_target is not None:
-            self.roi_target.setVisible(show)
 
 
 class ViewerDispatcher:
